working on run sim
run simulation
fixed type mismatch

+++ GamePhysics/src/Particle.cpp
@@ -15,12 +15,11 @@ Particle::Particle() {
 	mass = 0.0;
 }
 
-Particle::Particle(Vector2D pos, Vector2D v, float m) {
-	void setParticle(Vector2D pos, Vector2D v, float m);
-
+Particle::Particle(Vector2D pos, Vector2D v, double m) {
+	void setParticle(Vector2D pos, Vector2D v, double m);
 }
 
-void Particle::setParticle(Vector2D pos, Vector2D v, float m) {
+void Particle::setParticle(Vector2D pos, Vector2D v, double m) {
 	position = pos;
 	velocity = v;
 	mass = m;
@@ -32,7 +31,7 @@ void Particle::setPosition(Vector2D pos) {
 void Particle::setVelocity(Vector2D v) {
 	velocity = v;
 }
-void Particle::setMass(float m) {
+void Particle::setMass(double m) {
 	mass = m;
 }
 
@@ -44,7 +43,7 @@ Vector2D Particle::getVelocity() {
 	return velocity;
 }
 
-float Particle::getMass() {
+double Particle::getMass() {
 	return mass;
 }
